Tensorboard controller: Add scheduling functionality for Tensorboard servers that use RWO PVCs as log storages (kubeflow/kubeflow#5218)
* Add indexers as custom field selectors for list requests to cache

The tensorboard controller must be able to list pods that have
mounted a PVC with a specific ClaimName.

In order for this list request to cache to work properly, custom
field selectors are added. These selectors are used to index the
"pod.spec.volumes.persistentvolumeclaim.claimname" field so that
unneeded pods can be filtered out.

* Set pod's nodeAffinity if log files exist in a PVC

In the case of using a PVC as a logdir for Tensorboard Server, if
the PVC had a ReadWriteOnce access mode and was alread mounted by
another running pod X, then the Tensorboard Server pod would not
always be scheduled on the same node as X. As a result, the
Tensorboard Server pod would be blocked since multi-node access
is prohibited on ReadWriteOnce volumes.

In order for the Tensorboard Server pod to run successfully,
nodeAffinity was added to the spec.template.spec.affinity field
of the returned deployment.

As a result, both X and the Tensorboard
Server pod are now scheduled on the same node.

Resolves kubernetes/kubernetes#26567

* Set Tensorboard Server scheduling feature to 'off' by default

In the case that the Tensorboard Server used a RWO PVC (as a log
storage) that was already mounted by another pod, nodeAffinity
was used so that the Tensorboard Server would be scheduled
(if possible) on the same node as that pod.

Now, this added functionality is used only if the
'RWO_PVC_SCHEDULING' environmental variable is set to "true"
when running the Tensorboard controller.

This scheduling functionality is disabled by default.