jwa: Support for showing Notebooks from all namespaces (kubeflow/kubeflow#6706)
* jwa(back): Use gevent for gunicorn's workers

Since the backend only serves simple HTTP requests that need some K8s
data, instead of only have working processes we will also use greenlets.
This way we can have multiple user-space threads that will handle the
http requests.

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* jwa(front): Use cli to continuously build the UI

We extend the current package.json file to allow us to build and serve
the UI via Angular's dev server.

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* wa(front): Handle all-namespaces

The common library will add an event handler to dashboard's lib when it
is emitting all the namespaces. By extension, the common namespace
service will be exposing a new observable that will be emitting the list
of all the namespaces, when the user selects this from the dashboard.

This way current apps won't break, since they will simply not handle
this new observable.

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* wa(front): Handle dynamic table colums

The UIs will be showing a new column for the namespace only when the
user selects all-namespaces from the dashboard. The common code should
expose handlers for modifying the tables' config columns accordingly.

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* jwa(front): Make requests for all-namespaces

As a first iteration in order for the JWA to support showing Notebooks
from all namespaces it will be polling with batch requests. After all
the requests, for all the namespaces, succeed then the JWA will sort
them based on their name and show them in the table.

While this approach will be generating a significantly bigger amount of
requests, we chose this approach because:
1. It was simple to implement, since we only need to modify the frontend
2. We don't add any extra abstractions on the backend and keep the 1-1
   relationship with the k8s api endpoints
3. To further optimize we now only look on K8s concepts and won't have
   to adjust them to a custom backend implementation [2]

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>