web-apps: Ensure in-flight requests are cancelled  (kubeflow/kubeflow#6718)
* vwa(front): Use ng serve for local dev

Add proxy rules to connect the webpack dev server with the backend

* twa(front): Use ng serve for local dev

Refs arrikto/dev#1597

* vwa(back): Fix the install-deps make rule

Use the new common code path when installing the common packages.

* wa(front): Disable Angular analytics

Edit the angular.json file of all our frontend components to not send
Angular analytics. This way we won't be seeing a y/N input each time we
try to build the frontend, which could also block a CI/CD system.

* wa(front): Add exponential-backoff service

Create an Angular Service that handles the polling. Specifically the
polling function has as input an observable responsible for fetching the
data. It will return a final observable, which will be using the
fetch-observable, that will only be emitting the final data.

This service will take care of:
1. Checking if the new data is new and reset the polling
2. Cancelling in-flight requests, if someone terminates the subscription

* vwa(front): Use the new Polling Service

The frontend code of VWA will now be using the new Poller Service which
has a pure RxJS implementation underneath. This will make it simpler to
cancel in-flight requests and also moves the reset logic into the common
code.

* jwa(front): Use the new Polling Service

The frontend code of JWA will now be using the new Poller Service which
has a pure RxJS implementation underneath. This will make it simpler to
cancel in-flight requests and also moves the reset logic into the common
code.

* twa(front): Use the new Polling Service

The frontend code of TWA will now be using the new Poller Service which
has a pure RxJS implementation underneath. This will make it simpler to
cancel in-flight requests and also moves the reset logic into the common
code.